% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_borrow.R
\name{ec_borrow}
\alias{ec_borrow}
\title{Fisher Randomization Test and Conformal Selective Borrowing}
\usage{
ec_borrow(
  Y,
  A,
  S,
  X,
  method = "Conformal Selective Borrow AIPW",
  family = "gaussian",
  n_fisher = NULL,
  n_boot = NULL,
  gamma_sel = 0.6,
  cf = "cv+",
  cf_score = "CQR",
  cf_model = "glm",
  split_train = 0.75,
  cv_fold = 10,
  outcome_model = "glm",
  sampling_model = "glm",
  X_cw_ind = NULL,
  max_r = Inf,
  sig_level = 0.05,
  small_n_adj = TRUE,
  parallel = FALSE,
  n_cores = NULL,
  output_frt = FALSE
)
}
\arguments{
\item{Y}{Numeric vector of outcomes.}

\item{A}{Binary treatment indicator vector (1 for treatment, 0 for control).}

\item{S}{Binary indicator vector for source of data (1 for randomized trial
data, 0 for external control data).}

\item{X}{Matrix of covariates.}

\item{method}{Character string specifying the method for estimation. Options
include:
\itemize{
\item "No Borrow DiM"
\item "No Borrow AIPW"
\item "Borrow OM"
\item "Borrow IPW"
\item "Borrow staIPW"
\item "Borrow CW"
\item "Borrow AIPW"
\item "Borrow ACW"
\item "Conformal Selective Borrow AIPW" (the proposed & defalut)
\item "Conformal Selective Borrow ACW" (the proposed)
}}

\item{family}{A description of the outcome type. Default is "gaussian" for
continuous outcomes, but can also be "binomial" for binary outcomes.}

\item{n_fisher}{Integer specifying the number of randomizations for the
Fisher randomization test (FRT). Default is \code{NULL}, meaning FRT is skipped
to avoid long computation times without the user's permission. For reliable
FRT, set \code{n_fisher > 5000}.}

\item{n_boot}{Integer specifying the number of bootstrap samples for
computing standard errors (SE), only applicable for IPW/staIPW/CW/OM.
Default is \code{NULL}, meaning no bootstrap sampling is performed. To obtain
reliable SE, set \code{n_boot > 5000}.}

\item{gamma_sel}{Numeric value controlling the conformal selective borrowing
threshold. Default is \code{0.6}. It can also be adaptively determined using
\code{\link[=compute_ada_gamma]{compute_ada_gamma()}}.}

\item{cf}{Character string specifying the type of conformal inference to use.
Default is "cv+", with other options being "split", "full", and
"jackknife+".}

\item{cf_score}{Character string specifying the scoring function for
conformal inference. Default is "CQR" for conformal quantile regression.
Another option is "AR" for absolute residuals.}

\item{cf_model}{Character string specifying the model to use for outcome
prediction in conformal inference. Default is "glm" for generalized linear
model. Other options are "rf" for random forest and "ral" for relaxed
adaptive lasso.}

\item{split_train}{Numeric value (between 0 and 1) specifying the proportion
of data to use for training when \code{cf = "split"}. Default is 0.75.}

\item{cv_fold}{Integer specifying the number of folds for cross-validation
when \code{cf = "cv+"}. Default is 10.}

\item{outcome_model}{Character string specifying the model to use for outcome
imputation in OM/AIPW/ACW methods. Default is "glm" for generalized linear
model. Other options are "rf" for random forest and "ral" for relaxed
adaptive lasso.}

\item{sampling_model}{Character string specifying the model to use for
sampling model. Default is "glm" for generalized linear model.
Other options are "rf" for random forest and "ral" for relaxed adaptive
lasso.}

\item{X_cw_ind}{Integer vector of column indices in \code{X} to balance using
calibration weighting. Default is \code{NULL}: uses all columns, or applies
data-adaptive selection if \code{sampling_model = "ral"}.}

\item{max_r}{Numeric value limiting the ratio of residual variances between
trial and external control groups. Default is \code{Inf} (no restriction).}

\item{sig_level}{Numeric significance level for hypothesis tests. Default is
0.05.}

\item{small_n_adj}{Logical value indicating whether to adjust variance
estimation for small sample sizes in asymptotic normality inference. If
\code{TRUE}, the sandwich variance estimation will multiply by \code{n / (n - d)},
where \code{d} is the number of estimated parameters and \code{n} is the number of
units used to compute the sandwich variance estimator. Default is \code{TRUE}.}

\item{parallel}{Logical value indicating whether to use parallel computing.
Default is \code{FALSE}.}

\item{n_cores}{Integer specifying the number of cores to use for parallel
computing. Default is the number of available physical cores on the
machine, as determined by \code{future::availableCores(logical = FALSE)}.}

\item{output_frt}{Logical value indicating whether to output all test
statistic values from the Fisher randomization test. Default is \code{FALSE}.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{res}: A tibble containing estimation and inference results:
\itemize{
\item \code{est}: ATE estimate.
\item \code{se}: Standard error.
\item \code{ci_l}, \code{ci_u}: Confidence interval.
\item \code{p_value} (first row): P-value based on asymptotic normality.
\item \code{p_value} (second row): P-value based on FRT (If \code{n_fisher != NULL}).
\item \code{n_sel}: The number of borrowed external controls.
\item \code{ess_sel}: The effective sample size of borrowed external controls.
\item \code{runtime}: The computation time.
}
\item \code{id_sel}: The row numbers of selected external controls.
\item \code{dat_info}: A tibble containing main information about the original data:
\itemize{
\item \code{n_rt}: The number of units under randomized treatment.
\item \code{n_rc}: The number of units under randomized control.
\item \code{n_rct}: The total number of units in the randomized controlled trial.
\item \code{n_et}: The number of all external controls.
\item \code{id_et}: The row numbers of all external controls.
}
\item \code{gamma_sel}: Conformal selection threshold.
\item \code{out}: A tibble containing raw output.
\item \code{out_frt}: A tibble containing output of FRT, if \code{output_frt} is \code{TRUE}.
}
}
\description{
This function implements Fisher randomization tests (FRT) and Conformal
Selective Borrowing (CSB) in hybrid controlled trials. This function also
includes a series of external control borrowing estimators and their
inference results based on asymptotic normality for comparison (see the
\code{method} argument).
}
\examples{
# This example illustrates the use of the Fisher Randomization Test (FRT)
# and different borrowing methods for hybrid controlled trials.
\dontrun{
library(intFRT)

# Simulate data for a hybrid controlled trial
set.seed(1)
n_rct <- 50  # Number of observations in the randomized controlled trial
n_ec <- 100  # Number of external controls
n <- n_rct + n_ec  # Total number of observations

# Covariates (2 covariates, uniformly distributed)
X <- matrix(runif(n * 2), n, 2)

# Treatment assignment (1 = treatment, 0 = control)
A <- c(rep(0:1, each = n_rct / 2), rep(0, n_ec))

# Data source indicator (1 = randomized trial, 0 = external control)
S <- c(rep(1, n_rct), rep(0, n_ec))

# Generate potential outcomes (continuous)
Y1 <- 1 + 2 * X[,1] + 0.1 * X[,2] + rnorm(n)
Y0 <- 2 * X[,1] + 0.1 * X[,2] + rnorm(n)

# Introduce bias in half of the external controls
id_biased_EC <- tail(which(S == 0), n = n_ec / 2)
Y0[id_biased_EC] <- Y0[id_biased_EC] - 10

# Observed outcome
Y <- A * Y1 + (1 - A) * Y0

# Perform Fisher Randomization Test with No Borrowing
result_nb <- ec_borrow(
  Y = Y,
  A = A,
  S = S,
  X = X,
  method = "No Borrow AIPW",
  family = "gaussian",
  # FRT with 10 randomizations for illustration;
  # recommend n_fisher = 5000 for more powerful results
  n_fisher = 10
)

# View results
print(result_nb$res)

# View IDs of borrowed external controls (None for No Borrowing)
result_nb$id_sel

# Compute adaptive gamma (with a small n_rep_gamma for illustration)
ada_g <- compute_ada_gamma(
  Y, A, S, X,
  # Tuning with 20 replications for illustration purposes
  # Recommend setting `n_rep_gamma = 100` or higher with parallel computing
  n_rep_gamma = 20
)

# Perform Fisher Randomization Test with Conformal Selective Borrowing
result_csb <- ec_borrow(
  Y = Y,
  A = A,
  S = S,
  X = X,
  method = "Conformal Selective Borrow AIPW",
  family = "gaussian",
  gamma_sel = ada_g,
  # FRT with 10 randomizations for illustration
  # recommend n_fisher = 5000+ for more powerful results
  n_fisher = 10
)

# View results
print(result_csb$res)

# View IDs of borrowed external controls
result_csb$id_sel

# Mark borrowed external controls for plotting
sel <- rep(0, length(Y))
sel[result_csb$id_sel] <- 1

# Visualize borrowed external controls with ggplot
library(dplyr)
library(ggplot2)
library(forcats)
library(quantreg)

# Fit sampling score for multiple covariates
`Sampling Score` <- glm(S ~ X) \%>\% predict(type = "response")

# Prepare data for plotting
dat_plot <- tibble(Y, A, S, X, `Sampling Score`, sel) \%>\%
  mutate(
    Type = case_when(
      A == 1 & S == 1 ~ "RCT treated",
      A == 0 & S == 1 ~ "RCT control",
      A == 0 & S == 0 & sel == 1 ~ "External control (selected)",
      A == 0 & S == 0 & sel == 0 ~ "External control (unselected)"
    ) \%>\%
      as_factor \%>\%
      fct_relevel(
        "RCT treated", "RCT control",
        "External control (selected)", "External control (unselected)"
      )
  ) \%>\%
  filter(Type != "RCT treated")

# Fit quantile regressions to visualize the main range of RCT controls
fit975 <- rq(Y ~ `Sampling Score`, tau = 0.975, data = dat_plot,
             subset = dat_plot$Type == "RCT control")
dat_plot$pred975 <- predict(fit975, newdata = dat_plot)

fit025 <- rq(Y ~ `Sampling Score`, tau = 0.025, data = dat_plot,
             subset = dat_plot$Type == "RCT control")
dat_plot$pred025 <- predict(fit025, newdata = dat_plot)

# Plot results
dat_plot \%>\%
  ggplot() +
  geom_ribbon(aes(`Sampling Score`, ymin = pred025, ymax = pred975),
              fill = "grey80", alpha = 0.5) +
  geom_point(aes(`Sampling Score`, Y, color = Type)) +
  scale_color_manual(values = c("#5A5A5A", "#00ADFA", "#F8766D"))
}

}
